added allow filters and more changes (#32)
* added allow filters, changed NET displays, removed hubble and adjusted scub notifications

* some small fixes

* ran clippy and fmt

* fixed clippy complaints

* fixed deployment

+++ src/utils/reminders.rs
@@ -21,7 +21,7 @@ use crate::models::reminders::{
 
 pub async fn get_user_settings(db: &Database, id: u64) -> MongoResult<UserSettings> {
     db.collection("user_settings")
-        .find_one(doc! { "user": id }, None)
+        .find_one(doc! { "user": id as i64 }, None)
         .await?
         .ok_or_else(|| MongoError::from(MongoErrorKind::Io(Arc::new(IoErrorKind::NotFound.into()))))
         .and_then(|d| bson::from_document::<UserSettings>(d).map_err(|e| e.into()))
@@ -29,7 +29,7 @@ pub async fn get_user_settings(db: &Database, id: u64) -> MongoResult<UserSettin
 
 pub async fn get_guild_settings(db: &Database, id: u64) -> MongoResult<GuildSettings> {
     db.collection("guild_settings")
-        .find_one(doc! { "guild": id }, None)
+        .find_one(doc! { "guild": id as i64 }, None)
         .await?
         .ok_or_else(|| MongoError::from(MongoErrorKind::Io(Arc::new(IoErrorKind::NotFound.into()))))
         .and_then(|d| bson::from_document::<GuildSettings>(d).map_err(|e| e.into()))
